+++ scripts/actions/pf2e/pf2e-actions-pc.js
@@ -43,12 +43,14 @@ export class PcActionHandlerPf2e {
 
     /** @private */
     _forCharacter(result, tokenId, actor) {
+        let toggles = this._getTogglesCategory(actor, tokenId);
         let strikes = this._getStrikesList(actor, tokenId);
         let actions = this.baseHandler._getActionsList(actor, tokenId);
         let spells = this.baseHandler._getSpellsList(actor, tokenId);
         let feats = this.baseHandler._getFeatsList(actor, tokenId);
         let items = this.baseHandler._getItemsList(actor, tokenId);
         
+        this.baseHandler._combineCategoryWithList(result, this.i18n('tokenactionhud.toggles'), toggles);
         this.baseHandler._combineCategoryWithList(result, this.i18n('tokenactionhud.strikes'), strikes);
         this.baseHandler._combineCategoryWithList(result, this.i18n('tokenactionhud.actions'), actions);
         this.baseHandler._combineCategoryWithList(result, this.i18n('tokenactionhud.inventory'), items);
@@ -57,11 +59,24 @@ export class PcActionHandlerPf2e {
     }
 
     /** @private */
+    _getTogglesCategory(actor, tokenId) {
+        if (!settings.get('separateTogglesCategory'))
+            return;
+
+        let result = this.baseHandler.initializeEmptyCategory('toggles');
+        this._addTogglesCategories(actor, tokenId, result);
+
+        return result;
+    }
+
+    /** @private */
     _getStrikesList(actor, tokenId) {
         let result = this.baseHandler.initializeEmptyCategory('strikes');
         result.cssClass = 'oneLine';
         
-        this._addTogglesCategories(actor, tokenId, result);
+        if (!settings.get('separateTogglesCategory'))
+            this._addTogglesCategories(actor, tokenId, result);
+            
         this._addStrikesCategories(actor, tokenId, result);
 
         return result;
@@ -204,9 +219,11 @@ export class PcActionHandlerPf2e {
         let result = this.baseHandler.initializeEmptyCategory('attributes');
         let attributes = this.baseHandler.initializeEmptySubcategory();
 
-        let rollableAttributes = Object.entries(actor.data.data.attributes).filter(a => { if(a[1]) return !!a[1].roll });
+        let rollableAttributes = Object.entries(actor.data.data.attributes).filter(a => !!a[1]?.roll );
         let attributesMap = rollableAttributes.map(a => {
-            let name = a[0].charAt(0).toUpperCase()+a[0].slice(1);
+            let key = a[0];
+            let data = a[1];
+            let name = data.label ? data.label : key.charAt(0).toUpperCase()+key.slice(1);
             return { _id: a[0], name: name } 
         });