+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.9.25] 2020-11-27
+### Changed
+- PF2E - PC initiative label now reflects skill chosen
+- PF2E - Toggles by default go in a separate category but there is a setting to place them above strikes.
+
 ## [0.9.24] 2020-11-27
 ### Changed
 - Removed unnecessary setting for keeping categories open when clicked